feat: equational theorems of non-exposed defs should be private
This PR makes the equational theorems of non-exposed defs private. If
the author of a module chose not to expose the body of their function,
then they likely don't want that implementation to leak through
equational theorems. Helps with #8419.

With some luck doesn't even cause breakage for users outside the module
system (e.g. mathlib), as they'll just create their own copy of the
theorems if needed.