fix: let private names be unresolved in the pretty printer
This PR fixes an issue where private names are not unresolved when they are pretty printed, and an issue where in `pp.universes` mode names were allowed to shadow local names. Adds more delaboration tests for name unresolution.

It also cleans up the `delabConst` delaborator so that it uses `unresolveNameGlobalAvoidingLocals`, rather than doing any local context analysis itself. The `inPattern` logic has been removed; it appears to have been a heuristic added back in #575, but it's not clear what the current purpose is, and there are no tests that show that it is still needed.