fix: refine how `simp` tracks unfolded local definitions
This PR fixes a bug in `simp` where it was not resetting the set of zeta-delta reduced let definitions between `simp` calls. It also fixes a bug where `simp` would report zeta-delta reduced let definitions that weren't given as simp arguments (these extraneous let definitions appear due to certain processes temporarily setting `zetaDelta := true`). This PR also modifies the metaprogramming interface for the zeta-delta tracking functions so that they are re-entrant and prevent this kind of no-reset bug from occurring again. Closes #6655.

Re-entrance of this metaprogramming interface is not need to fix #6655, but it is needed for some future PRs.

The `tests/lean/run/6655.lean` file has an example of a deficiency of `simp?`, where `simp?` still over-reports unfolded let declarations.