feat: `debug.elabCheck` option to enable elaborator type checking
This PR adds an option called `debug.elabCheck` that when enabled applies the elaborator typechecker before sending declarations to the kernel. This is not a complete check. There is also a `debug.elabCheck.theoremValues` option to be able to turn off checking theorem bodies.

A purpose of `debug.elabCheck` is to make sure that theorem statements and definition types and values are type correct with respect to the elaborator.