feat: `let +generalize`
This PR adds the `+generalize` option to the `let` and `have` syntaxes. For example, `have +generalize n := a + b; body` replaces all instances of `a + b` in the expected type with `n` when elaborating `body`. This can be likened to a term version of the `generalize` tactic. One can combine this with `eq` in `have +generalize (eq := h) n := a + b; body` as an analogue to `generalize h : n = a + b`.